eliminar usuario y custom hook del dispatch
