+++ crud-ts-redux/src/types.d.ts
@@ -1,0 +1,12 @@
+export interface User {
+  avatar: string
+  name: string
+  email: string
+  github: string
+}
+
+export interface UserWithId extends User {
+  id: UserId
+}
+
+export type UserId = string
